[Feat] 추가 테마, 지역분류를 위한 기능 장소(식당, 관광지, 숙소를 조회하기 위한 regionKeyword, theme추가 및 관련 db 테이블과 관리코드 추가. 추후 여행지 저장기능 완성시 추가작업 필요. regionKeyword와 theme를 장소 튜플 생성과 동시에 areaCode, siGunGuCode, cat3를 사용해 조회하여 regionKeyword와 theme를 컬럼으로 추가하며 저장해야 함. 게시판 글 작업을 위한 지역 및 테마 분류 완성. 이를 이용해서 게시판 기능 추가 및 리팩토링 후 업데이트 예정.

+++ src/main/java/org/example/be/board/controller/BoardController.java
@@ -30,7 +30,7 @@ public class BoardController {
     }
     //게시판 글 조회
     @GetMapping("/{id}")
-    public ResponseEntity<CommonResponse<BoardDTO>> getBoardById(@PathVariable int id) {
+    public ResponseEntity<CommonResponse<BoardDTO>> getBoardById(@PathVariable Long id) {
         BoardDTO boardDTO = boardService.getBoard(id);
         return ResponseEntity.ok(CommonResponse.success(boardDTO, "게시글 조회 성공"));
     }
@@ -74,7 +74,7 @@ public class BoardController {
 
     // 게시글 삭제
     @DeleteMapping("/{id}")
-    public ResponseEntity<CommonResponse<String>> delete(@PathVariable int id) throws IOException {
+    public ResponseEntity<CommonResponse<String>> delete(@PathVariable Long id) throws IOException {
         boardService.deleteBoard(id);
         return ResponseEntity.status(HttpStatus.OK).body(CommonResponse.success(null,"게시글 삭제"));
     }
