+++ src/main/java/org/example/be/board/dto/CommentDTO.java
@@ -10,11 +10,11 @@ import java.time.LocalDateTime;
 @Setter
 @ToString
 public class CommentDTO {
-    private int id;
+    private Long id;
     private String commentWriter;
     private String commentContent;
-    private int boardId;
-    private Integer parentCommentId; // 대댓글에 사용
+    private Long boardId;
+    private Long parentCommentId; // 대댓글에 사용
     private LocalDateTime commentCreatedTime;
 
 }
