+++ src/main/java/org/example/be/board/service/CommentService.java
@@ -65,7 +65,7 @@ public class CommentService {
     }
 
     @Transactional
-    public void deletecomment(int id) {
+    public void deletecomment(Long id) {
         Comment comment = commentRepository.findById(id).orElseThrow(() -> new RuntimeException("댓글을 찾을 수 없습니다."));
         commentRepository.delete(comment);
         commentRepository.flush();
